+++ src/collections/Educations.ts
@@ -14,6 +14,11 @@ export const Educations: CollectionConfig = {
   admin: {
     defaultColumns: ['school', 'degree', 'startDate', 'endDate', 'updatedAt'],
     useAsTitle: 'school',
+    components: {
+      edit: {
+        editMenuItems: ['@/components/DeleteVersionsMenuItem#DeleteVersionsMenuItem'],
+      },
+    },
   },
   fields: [
     {